restructured from css to scss
