+++ client/package.json
@@ -13,6 +13,7 @@
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-scripts": "5.0.1",
+    "sass": "^1.65.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
