+++ client/public/index.html
@@ -25,6 +25,12 @@
       Learn how to configure a non-root public URL by running `npm run build`.
     -->
     <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@300;400;500;700;900&display=swap" rel="stylesheet">
+    <style>
+      * {
+        margin: 0;
+        font-family: 'Roboto', sans-serif;
+      }
+    </style>
     <title>React App</title>
   </head>
   <body>
