+++ client/src/App.jsx
@@ -1,6 +1,6 @@
 import logo from './logo.svg';
-import './App.css';
-import Home from './pages/Home';
+import './app.scss';
+import Home from './pages/home/Home';
 function App() {
   return (
     <div className="App">
