+++ client/src/components/featured/featured.scss
@@ -1,0 +1,90 @@
+.featured {
+    height: 90vh;
+    position: relative;
+    
+    .category {
+      position: absolute;
+      top: 80px;
+      left: 50px;
+      font-size: 30px;
+      font-weight: 500;
+      color: #fff;
+      display: flex;
+      align-items: center;
+      
+      select {
+        cursor: pointer;
+        background-color: var(--main-color);
+        border: 1px solid white;
+        color: #fff;
+        margin-left: 20px;
+        padding: 5px;
+      }
+    }
+    
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+    
+    .info {
+      width: 35%;
+      position: absolute;
+      left: 50px;
+      bottom: 100px;
+      color: #fff;
+      display: flex;
+      flex-direction: column;
+      
+      img {
+        width: 400px;
+      }
+      
+      .desc {
+        margin: 20px 0;
+      }
+      
+      .buttons {
+        display: flex;
+        
+        button {
+          padding: 10px 20px;
+          border: none;
+          border-radius: 5px;
+          display: flex;
+          align-items: center;
+          justify-content: center;
+          font-size: 18px;
+          font-weight: 500;
+          margin-right: 10px;
+          cursor: pointer;
+          
+          &.play {
+            background-color: #fff;
+            color: var(--main-color);
+          }
+          
+          &.more {
+            background-color: #808080;
+            color: #fff;
+          }
+          
+          span {
+            margin-left: 5px;
+          }
+        }
+      }
+    }
+  }
+  
+  
+  
+  
+  
+  
+  
+  
+  
+  
+
