+++ client/src/components/list/list.scss
@@ -1,0 +1,52 @@
+.list {
+    width: 100%;
+    margin-top: 10px;
+    
+    .list-title {
+      color: #fff;
+      font-size: 20px;
+      font-weight: 500;
+      margin-left: 50px;
+    }
+    
+    .wrapper {
+      position: relative;
+      .slider-arrow {
+        width: 50px;
+        height: 100%;
+        background-color: rgb(22, 22, 22, 0.5);
+        color: #fff;
+        position: absolute;
+        z-index: 99;
+        top: 0;
+        bottom: 0;
+        margin: auto;
+        cursor: pointer;
+        
+        &.left {
+          left: 0;
+        }
+        
+        &.right {
+          right: 0;
+        }
+        }
+        .list-container {
+          margin-left: 50px;
+          display: flex;
+          width: max-content;
+          margin-top: 10px;
+          transform: translateX(0px);
+          transition: all 1s ease;
+          /* overflow-x: auto; */
+        }
+    }
+}
+
+// .wrapper {
+//   position: relative;
+// }
+  
+  
+  
+
