+++ client/src/components/listItem/listItem.scss
@@ -1,0 +1,82 @@
+.list-item {
+    width: 225px;
+    height: 120px;
+    background-color: var(--main-color);
+    margin-right: 5px;
+    overflow: hidden;
+    cursor: pointer;
+    color: #fff;
+    
+    img {
+      width: 100%;
+      height: 100%;
+      object-fit: cover;
+    }
+    
+    video {
+      width: 100%;
+      height: 140px;
+      object-fit: cover;
+      position: absolute;
+      top: 0;
+      left: 0;
+    }
+    
+    &:hover {
+      width: 325px;
+      height: 300px;
+      position: absolute;
+      top: -150px;
+      -webkit-box-shadow: 0px 0px 15px 0px rgba(255,255,255,0.07); 
+      box-shadow: 0px 0px 15px 0px rgba(255,255,255,0.07);
+      border-radius: 5px;
+      
+      img {
+        height: 140px;
+      }
+  
+      .item-info {
+        display: flex;
+        flex-direction: column;
+        padding: 5px;
+        
+        .list-icons {
+          display: flex;
+          margin-bottom: 10px;
+          
+          .info-icon {
+            border: 2px solid #fff;
+            padding: 5px;
+            border-radius: 50%;
+            margin-right: 10px;
+            font-size: 16px;
+          }
+        }
+  
+        .item-info-top {
+          display: flex;
+          align-items: center;
+          margin-bottom: 10px;
+          font-size: 14px;
+          font-weight: 600;
+          color: #808080;
+          
+          .limit {
+            border: 1px solid #808080;
+            padding: 1px 3px;
+            margin: 0 10px;
+          }
+        }
+  
+        .info-description {
+          font-size: 13px;
+          margin-bottom: 10px;
+        }
+        
+        .genre {
+          font-size: 14px;
+          color: #d3d3d3
+        }
+      }
+    }
+  }
