+++ client/src/pages/home/home.scss
@@ -1,0 +1,4 @@
+.home {
+    background-color: var(--main-color);
+    overflow: hidden;
+  }
